refactor: 피드 리스트, 아이템 공용 컴포넌트로 개선
